requirements.txt
htmllistparse==0.5.2
Pillow==6.2.1
